alter Float32 test

+++ test/test_randoms.jl
@@ -42,7 +42,7 @@
 
         rspure_float32 = randstate(AbstractArray{Float32}, nmodes, pure = true)
         @test all(i -> ((i >= 0) || isapprox(i, 0.0, atol = 1e-3)), real(eigvals(rspure_float32.covar .+ im*Omega)))
-        @test isapprox(purity(rspure_float32), 1.0, atol = 1e-3)
+        @test isapprox(Float64(purity(rspure_float32)), 1.0, atol = 1e-3)
 
         rs_static = randstate(SVector{2*nmodes}, SMatrix{2*nmodes,2*nmodes}, nmodes)
         rc_static = randchannel(SVector{2*nmodes}, SMatrix{2*nmodes,2*nmodes}, nmodes)
